+++ tests/test_configuration.py
@@ -534,3 +534,10 @@ def test_get_vendor_id(motion_controller):
     vendor_id = mc.configuration.get_vendor_id(alias)
 
     assert vendor_id == expected_vendor_id
+
+
+@pytest.mark.no_connection
+def test_change_node_id_exception(motion_controller):
+    mc, alias = motion_controller
+    with pytest.raises(ValueError):
+        mc.configuration.change_node_id(32, alias)